+++ test/base/api.js
@@ -65,8 +65,8 @@ test("correct api should be exposed", function() {
             "values",
             "entries",
             "when",
-            "startAction",
-            "endAction"
+            "_startAction",
+            "_endAction"
         ].sort()
     )
 })